+++ bedwars-core/src/main/java/fr/heneria/bedwars/core/map/operation/MapOperationType.java
@@ -3,5 +3,6 @@ package fr.heneria.bedwars.core.map.operation;
 public enum MapOperationType {
   DUPLICATE,
   BACKUP,
+  IMPORT,
   DELETE
 }